# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MovieApp",
        "description": "Api para vizualização de informações de filmes migrados a partir de uma dataset",
        "contact": {
            "name": "Vitor Guedes",
            "email": "vtrf2.0@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://dev.backend.com",
            "description": "Ambiente local",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http",
                    "description": "Protocolo usado"
                }
            }
        }
    ],
    "paths": {
        "/v1/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Dados do usuário logado",
                "operationId": "d2e4c8fd5e403bcfa95c87cb0fcf893d",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cria um novo usuário no sistema",
                "operationId": "9ec4eb6039180ebf8b70bc9d02d909af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usário",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email do usuário usado para obeter o token de acesso",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha de acesso do usuário",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmação da senha de acesso do usuário",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso"
                    },
                    "422": {
                        "description": "Parametros inválidos ou nulos"
                    }
                }
            }
        },
        "/v1/api/users/token": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtem o token de acesso da api",
                "operationId": "5b80c6f721f6b7ecd8d4da953db41dd7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email de acesso",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha de acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token de acesso",
                                            "type": "string"
                                        },
                                        "type": {
                                            "description": "Tipo de token",
                                            "type": "string"
                                        },
                                        "expire": {
                                            "description": "Tempo de expiração",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Credênciais inválidas"
                    }
                }
            }
        },
        "/v1/api/users/logout": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Inválidar token de acesso",
                "operationId": "aaf7ea2371d78233e2834730c7ba011d",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "description": "Token de acesso",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}